+++ packages/workdo/Taskly/src/Http/Requests/UpdateProjectContractRequest.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Workdo\Taskly\Http\Requests;
+
+class UpdateProjectContractRequest extends StoreProjectContractRequest
+{
+}